requirements.txt
pyyaml~=6.0.1
torch~=2.0.1
torchvision~=0.15.2
numpy~=1.24.4
Pillow~=10.0.0
tqdm~=4.66.1
pandas
gdown
tensorboard